rewritten according to Andrzej's suggestions :-)
