+++ C_bool.BLL/Models/Users/User.cs
@@ -46,33 +46,6 @@ namespace C_bool.BLL.Models.Users
 
         public int Points { get; set; }
 
-        /// <summary>
-        /// prints basic information about user based on Id, Id might be empty
-        /// </summary>
-        /// <param name="list">input List of User objects</param>
-        /// <param name="id">Id number of user, if empy - prints all users</param>
-        public static void PrintInformation(List<User> list, string id)
-        {
-            foreach (var user in list)
-            {
-                var outputString = $"\t| Imię: {user.FirstName}\n\t| Nazwisko: {user.LastName}\n\t| Płeć: {user.Gender}\n\t| Wiek: {user.Age}\n\t| Adres: {user.Address}\n\t| E-mail: {user.Email}\n\t| Telefon: {user.Phone}\n\t| Firma: {user.Company}\n\t------------\n\t| Aktywny: {user.IsActive}\n\t| Szer. geo.: {user.Latitude}\n\t| Wys. geo.: {user.Longitude}\n\t| Punkty: {user.Points}\n";
 
-                if (user.Id.Equals(id))
-                {
-                    Console.ForegroundColor = ConsoleColor.Yellow;
-                    Console.WriteLine($"ID: {user.Id}");
-                    Console.ResetColor();
-                    Console.Write(outputString);
-                    return;
-                }
-                else if (id.Length == 0)
-                {
-                    Console.ForegroundColor = ConsoleColor.Yellow;
-                    Console.WriteLine($"ID: {user.Id}");
-                    Console.ResetColor();
-                    Console.Write(outputString);
-                }
-            }
-        }
     }
 }
